backend/requirements.txt
autogen-agentchat
autogen-core
autogen-ext[openai,azure]
fastapi
uvicorn
pydantic
python-multipart
sse-starlette
python-dotenv
aiofiles
chardet
jieba
PyPDF2
python-docx
openpyxl
sentence-transformers
chromadb
langchain-community
jinja2
markdown
beautifulsoup4
requests
httpx
pymysql
sqlalchemy